Fix to make to display connected applications unique
