--- app/views/apps/index.html+mobile.slim
@@ -1,9 +1,0 @@
-- title = t "head.title.apps.index"
-- set_meta_tags(title: title, reverse: true, prefix: false)
-- set_page_category "apps_index"
-
-.c-card.p-1.mb-2
-  = render "apps/app_list", access_tokens: @access_tokens
-
-= render "application/components/footer"
-= render "application/components/tabbar"
