--- app/views/apps/index.html.slim
@@ -1,12 +1,0 @@
-- title = t "head.title.apps.index"
-- set_meta_tags(title: title, reverse: true, prefix: false)
-- set_page_category "apps_index"
-
-= render "application/components/navbar"
-= render "settings/subnav"
-
-.p-application__content
-  .c-card.u-mw-600.mx-auto.mt-1.mb-2.p-1
-    = render "apps/app_list", access_tokens: @access_tokens
-
-= render "application/components/footer"
