+++ app/views/settings/apps/index.html.slim
@@ -1,0 +1,13 @@
+- title = t "head.title.settings.apps.index"
+- set_meta_tags(title: title, reverse: true, prefix: false)
+- set_page_category "settings_apps_index"
+
+= render "application/components/navbar"
+= render "settings/subnav"
+
+.p-application__content
+  h2.h4.text-center.my-2.font-weight-bold = t "noun.applications"
+  .c-card.u-mw-600.mx-auto.p-1.mt-1.mb-2
+    = render "settings/apps/app_list", apps: @apps
+
+= render "application/components/footer"
