+++ app/views/settings/index.html+mobile.slim
@@ -13,9 +13,9 @@
         = icon "user", class: "mr-zp5 text-muted"
         = t "head.title.accounts.show"
     li.mb-zp5
-      = link_to apps_path do
+      = link_to settings_apps_path do
         = icon "plug", class: "mr-zp5 text-muted"
-        = t "head.title.apps.index"
+        = t "head.title.settings.apps.index"
     li.mb-zp5
       = link_to providers_path do
         = icon "share-alt", class: "mr-zp5 text-muted"
