+++ config/locales/noun.en.yml
@@ -11,6 +11,7 @@ en:
     annict_db: Annict DB
     annict_forum: Annict Forum
     annict: Annict
+    applications: Applications
     basic_info: Basic Info
     before: Before
     bio: Bio
