+++ config/locales/noun.ja.yml
@@ -11,6 +11,7 @@ ja:
     annict_db: Annict DB
     annict_forum: Annict Forum
     annict: Annict
+    applications: アプリケーション
     basic_info: 基本情報
     before: Before
     bio: プロフィール
