+++ config/locales/verb.en.yml
@@ -12,6 +12,7 @@ en:
     next: Next
     post: Post
     previous: Prev
+    revoke: Revoke
     search: Search
     sign_out: Sign Out
     skip: Skip
