+++ config/locales/verb.ja.yml
@@ -12,6 +12,7 @@ ja:
     next: 次へ
     post: 投稿する
     previous: 前へ
+    revoke: 接続解除する
     search: 検索する
     sign_out: ログアウト
     skip: スキップする
